+++ src/main/java/com/gmail/val59000mc/configuration/MainConfig.java
@@ -132,6 +132,7 @@ public class MainConfig extends YamlFile {
 	public static final BasicOption<Boolean> END_GAME_WHEN_ALL_PLAYERS_HAVE_LEFT = new BasicOption<>("countdown-ending-game-when-all-players-have-left",true);
 	public static final BasicOption<Boolean> DEBUG = new BasicOption<>("debug",false);
 	public static final BasicOption<Boolean> ONE_PLAYER_MODE = new BasicOption<>("one-player-mode",false);
+	public static final BasicOption<String> LOGGING_LEVEL = new BasicOption<>("logging-level", "INFO");
 
 	// Pre-generate world
 	public static final BasicOption<Boolean> ENABLE_PRE_GENERATE_WORLD = new BasicOption<>("pre-generate-world.enable",true);